% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair2dist.R
\name{pair2dist}
\alias{pair2dist}
\title{Helper function to transform long matrix to distance matrix}
\usage{
pair2dist(dat, colname)
}
\arguments{
\item{dat}{Long matrix with three columns (site1, site2 and values)}

\item{colname}{Name of the column that contains distance value}
}
\value{
Object of class \code{dist}
}
\description{
Helper function to transform long matrix to distance matrix
}
