% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_c_a_pd.R
\name{pair_c_a_pd}
\alias{pair_c_a_pd}
\title{Compute a single phylogenetic dissimilarity estimates between two sites}
\usage{
pair_c_a_pd(pair_cols, com, level, PDtree)
}
\arguments{
\item{pair_cols}{A vector of two column names representing the pair of sites.}

\item{com}{A community data matrix (data.frame or matrix) with sites as columns and species as rows.}

\item{level}{A numeric value indicating the coverage level for diversity estimation.}

\item{PDtree}{A phylogenetic tree object of class 'phylo'.}
}
\value{
A list containing site names and dissimilarity estimates (Sørensen, Horn, Morisita-Horn).
}
\description{
Compute a single phylogenetic dissimilarity estimates between two sites
}
\details{
PDtype A character string indicating the type of phylogenetic diversity to compute (currently hard coded as "meanPD").
}
\references{
Chao, A., Thorn, S., Chiu, C.-H., Moyes, F., Hu, K.-H., Chazdon, R. L., Wu, J., Magnago, L. F. S., Dornelas, M., Zeleny, D., Colwell, R. K., and Magurran, A. E. (2023). Rarefaction and extrapolation with beta diversity under a framework of Hill numbers: the iNEXT.beta3D standardization. Ecological Monographs e1588.
}
