% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_c_a_td.R
\name{pair_c_a_td}
\alias{pair_c_a_td}
\title{Compute a single pairwise similarity estimates based on TD diversity}
\usage{
pair_c_a_td(pair_cols, com, level)
}
\arguments{
\item{pair_cols}{A vector of two column names representing the pair of sites.}

\item{com}{A community data matrix (data.frame or matrix) with sites as columns and species as rows.}

\item{level}{A numeric value indicating the coverage level for diversity estimation.}
}
\value{
A list containing site names and similarity estimates (Sørensen, Horn, Morisita-Horn).
}
\description{
Compute a single pairwise similarity estimates based on TD diversity
}
\details{
This function computes pairwise similarity estimates between two sites using TD diversity measures.
It calculates gamma and alpha diversities and derives dissimilarity indices based on these values.
}
\references{
Chao, A., Thorn, S., Chiu, C.-H., Moyes, F., Hu, K.-H., Chazdon, R. L., Wu, J., Magnago, L. F. S., Dornelas, M., Zeleny, D., Colwell, R. K., and Magurran, A. E. (2023). Rarefaction and extrapolation with beta diversity under a framework of Hill numbers: the iNEXT.beta3D standardization. Ecological Monographs e1588.
}
