% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{safe_get}
\alias{safe_get}
\title{Helper functions}
\usage{
safe_get(fun, pkg)
}
\arguments{
\item{fun}{Function to be loaded from a package function that is not exported from that package (not in NAMESPACE)}

\item{pkg}{R package from which to load the function from}
}
\description{
Helper functions
}
\details{
Workaround proposed here \link{https://stackoverflow.com/questions/63023526/unexported-object-imported-by-a-call-tsfeaturesscalets}
}
\keyword{internal}
