% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pairwise.R
\name{run_pairwise}
\alias{run_pairwise}
\title{Run pairwise computations in parallel using forked processes}
\usage{
run_pairwise(
  com,
  fun,
  level = 0.8,
  ncores = 12,
  PDtree = NULL,
  chunk_size = 5000
)
}
\arguments{
\item{com}{A community matrix (samples x species)}

\item{fun}{A function to apply to each pair of species}

\item{level}{A numeric value indicating the level parameter for the function}

\item{ncores}{Number of cores to use for parallel processing}

\item{PDtree}{An optional phylogenetic tree object}

\item{chunk_size}{Number of pairs to process in each chunk}
}
\value{
A data.table with the results of the pairwise computations
}
\description{
Run pairwise computations in parallel using forked processes
}
\details{
Only use with ncores > 1.
}
\references{
Chao, A., Thorn, S., Chiu, C.-H., Moyes, F., Hu, K.-H., Chazdon, R. L., Wu, J., Magnago, L. F. S., Dornelas, M., Zeleny, D., Colwell, R. K., and Magurran, A. E. (2023). Rarefaction and extrapolation with beta diversity under a framework of Hill numbers: the iNEXT.beta3D standardization. Ecological Monographs e1588.
}
